==> notebooks/01 - Data Collection.ipynb <==
# Collect and generate data

### Note
- Originally this notebook contains series of commands to download data from kaggle and UN comtrade. 
- I also considered creating synthetic dataset using Faker library
- Ultimately, I just decided to create a new account in UN comtrade and downloaded bulk data

import pandas as pd
import numpy as np

# remove column restrictions
pd.options.display.max_columns = None

# scan raw data from ../data/raw

raw_data_path = '../data/raw'

# List all files in the raw data directory
import os
raw_files = os.listdir(raw_data_path)
print("Raw data files:", raw_files)

for file in raw_files:
    file_path = os.path.join(raw_data_path, file)
    df = pd.read_csv(file_path, encoding="latin-1", engine="python")
    print(f"Data from {file}:")
    display(df.head())  # Display the first few rows of each file
    print("\n")
    print(f"Shape of {file}: {df.shape}")

# get a sample of this raw trade data
sample_data = df.sample(n=5, random_state=42)
display(sample_data)

# save to csv
sample_data.to_csv('../data/processed/sample_trade_data.csv', index=False)
