==> notebooks/03 - NL to pandas (SCRAPPED).ipynb <==
import pandas as pd 
import os
import openai
import json

# remove column restrictions
pd.set_option('display.max_columns', None)

# supress scientific notation
pd.set_option('display.float_format', lambda x: '%.3f' % x)

### Load parquet into df

processed_data_path = '../data/processed/trade_data_cleaned.parquet'

print(f'Loading data from {processed_data_path}')
df = pd.read_parquet(processed_data_path)

print("Columns in the DataFrame:")
print(df.columns.tolist())
print(f'Data loaded with shape: {df.shape}')
display(df.head())

### Define system prompt

SYSTEM_PROMPT = """ 
You are a trade data analyst.
Return ONLY valid JSON with these keys:
filter   – boolean condition in plain English
  groupby  – list of columns
  agg      – {column: aggfunc}
  stats    – {stat_name: [columns]}       # e.g. "corr": ["value","netwgt"], or "describe": ["value"]
  sort     – optional {by: col, ascending: bool}
  top_n    – optional int
  chart    – "bar" | "line" | null

  Columns you may reference:
  shipper, consignee, description,
  value, period, reporterdesc, partnerdesc.
"""


from openai import OpenAI
client = OpenAI()

question = "What is the total value of shipments by shipper in 2023?"

resp = client.chat.completions.create(
    model="gpt-4",
    messages=[
        {"role": "system", "content": SYSTEM_PROMPT},
        {"role": "user", "content": question}
    ],
    temperature=0.0,
    )

plan_text = resp.choices[0].message.content
print(f"Plan text: {plan_text}")

# parse to dict
plan = json.loads(plan_text)
print(f"Plan dict: {plan}")

question = "Show me the correlation between value and netwgt for 2022."
resp = client.chat.completions.create(
    model="gpt-4",
    messages=[
        {"role": "system", "content": SYSTEM_PROMPT},
        {"role": "user", "content": question}
    ],
    temperature=0.0,
)
plan_text = resp.choices[0].message.content
print(f"Plan text: {plan_text}")

# parse to dict
plan = json.loads(plan_text)
print(f"Plan dict: {plan}")

### Test utils

from app.utils import answer

question = "Which country has the highest total value of shipments in 2023?"
out = answer(question= question)


out['df']
